+++ R/covid19br.R
@@ -1,5 +1,6 @@
 #' Brazilian COVID-19 Pandemic Data.
 #'
+#' @import httr2 sf tibble
 #' @importFrom rlang .data
 #' @importFrom dplyr %>%
 #' @importFrom dplyr as_tibble
@@ -24,7 +25,7 @@
 #' @description The package provides a function to automatically import  Brazilian CODID-19 pandemic data into R. Brazilian data is available on the country, region, state, and city levels, and are obtained from the official Brazilian repository at <https://covid.saude.gov.br/>. The package also downloads the world-level COVID-19 data from the John Hopkins University's repository at <https://github.com/CSSEGISandData/COVID-19>.
 #'
 #'@author Fábio N. Demarqui, Cristiano C. Santos, and Matheus B. Costa.
-#' @docType package
+#' _PACKAGE
 #' @name covid19br
 #' @aliases covid19br-package