+++ NEWS.md
@@ -50,3 +50,8 @@
 # covid19br 0.1.8
 
 - All vignettes has been removed.
+
+# covid19br 1.0.0
+
+- Download function was completely rewriten.
+- COVID-19 data is available until May 5, 2023, and will no longer be updated.